requirements.txt
click==4.0
numpy==1.9.2
pandas==0.16.2
PyInstaller==2.1
python-dateutil==2.4.2
pytz==2015.4
pywin32==219
six==1.9.0
wheel==0.24.0
XlsxWriter==0.7.3
